+++ src/integrations.ts
@@ -1,4 +1,4 @@
-import { Integration, IntegrationClass } from '@sentry/types';
+import { Integration } from '@sentry/types';
 import { dynamicRequire } from '@sentry/utils';
 import { ElectronOptions } from './';
 
@@ -8,7 +8,6 @@ import {
   ElectronMinidump,
   OnUncaughtException,
   PreloadInjection,
-  RendererIPC,
   SentryMinidump,
   MainProcessSession,
 } from './main/integrations';
@@ -17,18 +16,17 @@ import { EventToMain, RendererContext, ScopeToMain } from './renderer/integratio
 /** Convenience interface used to expose Integrations */
 export interface Integrations {
   // For main process
-  SentryMinidump: IntegrationClass<SentryMinidump>;
-  ElectronMinidump: IntegrationClass<ElectronMinidump>;
-  ElectronEvents: IntegrationClass<ElectronEvents>;
-  MainContext: IntegrationClass<MainContext>;
-  RendererIPC: IntegrationClass<RendererIPC>;
-  OnUncaughtExcept: IntegrationClass<OnUncaughtException>;
-  PreloadInjection: IntegrationClass<PreloadInjection>;
-  MainProcessSession: IntegrationClass<MainProcessSession>;
+  SentryMinidump: SentryMinidump;
+  ElectronMinidump: ElectronMinidump;
+  ElectronEvents: ElectronEvents;
+  MainContext: MainContext;
+  OnUncaughtExcept: OnUncaughtException;
+  PreloadInjection: PreloadInjection;
+  MainProcessSession: MainProcessSession;
   // For renderer process
-  ScopeToMain: IntegrationClass<ScopeToMain>;
-  EventToMain: IntegrationClass<EventToMain>;
-  RendererContext: IntegrationClass<RendererContext>;
+  ScopeToMain: ScopeToMain;
+  EventToMain: EventToMain;
+  RendererContext: RendererContext;
 }
 
 /** Return all Electron integrations and add EmptyIntegrations for integrations missing in this process. */
@@ -48,7 +46,6 @@ export function getIntegrations(): Integrations {
         ElectronMinidump: EmptyIntegration,
         ElectronEvents: EmptyIntegration,
         MainContext: EmptyIntegration,
-        RendererIPC: EmptyIntegration,
         OnUncaughtExcept: EmptyIntegration,
         PreloadInjection: EmptyIntegration,
         MainProcessSession: EmptyIntegration,
@@ -85,13 +82,13 @@ class EmptyIntegration implements Integration {
 /** Filters out any EmptyIntegrations that are found */
 export function removeEmptyIntegrations(options: ElectronOptions): void {
   if (Array.isArray(options.integrations)) {
-    options.integrations = options.integrations.filter((i) => i.name !== 'EmptyIntegration');
+    options.integrations = options.integrations.filter((i) => i.name !== EmptyIntegration.id);
   } else if (typeof options.integrations === 'function') {
     const originalFn = options.integrations;
 
     options.integrations = (integrations) => {
       const userIntegrations = originalFn(integrations);
-      return userIntegrations.filter((integration) => integration.name !== 'EmptyIntegration');
+      return userIntegrations.filter((integration) => integration.name !== EmptyIntegration.id);
     };
   }
 }