+++ src/main/context.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-lines */
 import { Event, SdkInfo } from '@sentry/types';
 import * as child from 'child_process';
 import { app } from 'electron';
@@ -5,9 +6,9 @@ import { platform, release } from 'os';
 import { join } from 'path';
 
 import { readDirAsync, readFileAsync } from './fs';
+import { SDK_VERSION } from './version';
 
 export const SDK_NAME = 'sentry.javascript.electron';
-export const SDK_VERSION = '3.0.0-beta';
 
 /** Operating system context information. */
 interface OsContext {
@@ -276,6 +277,12 @@ async function _getEventDefaults(release?: string): Promise<Event> {
     environment: process.defaultApp ? 'development' : 'production',
     release: release || `${name.replace(/\W/g, '-')}@${app.getVersion()}`,
     user: { ip_address: '{{auto}}' },
+    tags: {
+      'event.origin': 'electron',
+      'event.environment': 'javascript',
+      // Legacy way of filtering native vs JavaScript events
+      event_type: 'javascript',
+    },
   };
 }