+++ src/main/integrations/electron-minidump.ts
@@ -1,12 +1,13 @@
 import { getCurrentHub, Scope } from '@sentry/core';
-import { NodeClient } from '@sentry/node';
-import { Event, Integration, ScopeContext } from '@sentry/types';
+import { NodeClient, NodeOptions } from '@sentry/node';
+import { Event, Integration } from '@sentry/types';
 import { Dsn, forget, logger, SentryError } from '@sentry/utils';
 import { app, crashReporter } from 'electron';
 
 import { mergeEvents, normalizeEvent } from '../../common';
 import { getEventDefaults } from '../context';
-import { rendererRequiresCrashReporterStart, usesCrashpad } from '../electron-normalize';
+import { onRendererProcessGone, rendererRequiresCrashReporterStart, usesCrashpad } from '../electron-normalize';
+import { sessionCrashed } from './main-process-session';
 
 /** Is object defined and has keys */
 function hasKeys(obj: any): boolean {
@@ -14,7 +15,7 @@ function hasKeys(obj: any): boolean {
 }
 
 /** Gets a Scope object with user, tags and extra */
-function getScope(): Partial<ScopeContext> {
+function getScope(options: NodeOptions): Event {
   const scope = getCurrentHub().getScope() as any | undefined;
 
   if (!scope) {
@@ -22,6 +23,8 @@ function getScope(): Partial<ScopeContext> {
   }
 
   return {
+    release: options.release,
+    environment: options.environment,
     /* eslint-disable @typescript-eslint/no-unsafe-member-access */
     ...(hasKeys(scope._user) && { user: scope._user }),
     ...(hasKeys(scope._tags) && { tags: scope._tags }),
@@ -65,13 +68,18 @@ export class ElectronMinidump implements Integration {
       throw new SentryError(`The '${this.name}' integration is only supported with Electron >= v9`);
     }
 
-    const dsnString = getCurrentHub().getClient<NodeClient>()?.getOptions()?.dsn;
+    const options = getCurrentHub().getClient<NodeClient>()?.getOptions();
 
-    if (!dsnString) {
+    if (!options || !options?.dsn) {
       throw new SentryError('Attempted to enable Electron native crash reporter but no DSN was supplied');
     }
 
-    this._startCrashReporter(dsnString);
+    this._startCrashReporter(options);
+
+    // If a renderer process crashes, mark any existing session as crashed
+    onRendererProcessGone((_, __) => {
+      sessionCrashed();
+    });
 
     // If we're using the Crashpad minidump uploader, we set extra parameters whenever the scope updates
     if (usesCrashpad()) {
@@ -82,10 +90,10 @@ export class ElectronMinidump implements Integration {
   /**
    * Starts the native crash reporter
    */
-  private _startCrashReporter(dsn: string): void {
+  private _startCrashReporter(options: NodeOptions): void {
     // We don't add globalExtra when Breakpad is in use because it doesn't support JSON like strings:
     // https://github.com/electron/electron/issues/29711
-    const globalExtra = usesCrashpad() ? { sentry___initialScope: JSON.stringify(getScope()) } : undefined;
+    const globalExtra = usesCrashpad() ? { sentry___initialScope: JSON.stringify(getScope(options)) } : undefined;
 
     logger.log('Starting Electron crashReporter');
 
@@ -93,7 +101,7 @@ export class ElectronMinidump implements Integration {
       companyName: '',
       ignoreSystemCrashHandler: true,
       productName: app.name || app.getName(),
-      submitURL: minidumpUrlFromDsn(new Dsn(dsn)),
+      submitURL: minidumpUrlFromDsn(new Dsn(options.dsn || '')),
       uploadToServer: true,
       compress: true,
       globalExtra,
@@ -139,7 +147,7 @@ export class ElectronMinidump implements Integration {
   /** Builds up an event to send with the native Electron uploader */
   private async _getNativeUploaderEvent(scope: Scope): Promise<Event> {
     const event = mergeEvents(await getEventDefaults(), {
-      tags: { event_type: 'native' },
+      tags: { 'event.environment': 'native', event_type: 'native' },
     });
 
     // Apply the scope to the event