+++ src/renderer/sdk.ts
@@ -13,9 +13,20 @@ export const defaultIntegrations = [...defaultBrowserIntegrations, new ScopeToMa
  * Initialize Sentry in the Electron renderer process
  */
 export function init(options: BrowserOptions): void {
+  // eslint-disable-next-line no-restricted-globals
   if (window.__SENTRY_IPC__ === undefined) {
-    throw new SentryError(`Communication with the Electron main process could not be established
-See the docs: https://docs.sentry.io/platforms/javascript/guides/electron/#preload`);
+    throw new SentryError(`Communication with the Electron main process could not be established.
+
+This is likely because the preload script was not run.
+Preload scripts are usually injected automatically but this can fail if you are bundling the Electron main process code.
+
+The required preload code can be imported via:
+  require('@sentry/electron/preload');
+or
+  import '@sentry/electron/preload';
+
+Check out the Webpack test app for an example of how to configure this:
+https://github.com/getsentry/sentry-electron/blob/master/test/e2e/test-apps/isolated-app`);
   }
 
   // We don't want browser session tracking enabled by default because we already have Electron
@@ -31,11 +42,6 @@ See the docs: https://docs.sentry.io/platforms/javascript/guides/electron/#prelo
   if (options.dsn === undefined) {
     // Events are sent via the main process but browser SDK wont start without dsn
     options.dsn = 'https://12345@dummy.dsn/12345';
-  }
-
-  if (options.release === undefined) {
-    // Events are sent via the main process but tracing wont work without release
-    options.release = 'fake-release';
   }
 
   // We only handle initialScope in the main process otherwise it can cause race conditions over IPC