+++ test/e2e/server.ts
@@ -12,15 +12,17 @@ export interface TestServerEvent<T = unknown> {
   /** Request ID (UUID) */
   appId: string;
   /** Public auth key from the DSN. */
-  sentryKey: string;
+  sentry_key: string;
+  /** Sentry Event data */
+  eventData?: Event;
+  /** Sentry Sessions data */
+  sessionData?: Session;
   /** Extra namespaced form data */
   namespacedData?: { [key: string]: any };
   /** An optional minidump file, if included in the event. */
   dumpFile?: boolean;
   /** API method used for submission */
   method: 'envelope' | 'minidump' | 'store';
-  /** Sentry Event data */
-  data: T;
 }
 
 /**
@@ -66,18 +68,18 @@ export class TestServer {
 
       if (type === 'session') {
         this._addEvent({
-          data: JSON.parse(payload) as Session,
-          dumpFile: false,
-          appId: ctx.params.id,
-          sentryKey: keyMatch[1],
+          sessionData: JSON.parse(payload) as Session,
+          dump_file: false,
+          id: ctx.params.id,
+          sentry_key: keyMatch[1],
           method: `envelope`,
         });
       } else {
         this._addEvent({
-          data: JSON.parse(payload) as Event,
-          dumpFile: !!attachmentHeaders && !!attachment && attachment.startsWith('MDMP'),
-          appId: ctx.params.id,
-          sentryKey: keyMatch[1],
+          eventData: JSON.parse(payload) as Event,
+          dump_file: !!attachmentHeaders && !!attachment && attachment.startsWith('MDMP'),
+          id: ctx.params.id,
+          sentry_key: keyMatch[1],
           method: `envelope`,
         });
       }
@@ -101,11 +103,11 @@ export class TestServer {
         const dump_file = result.files.upload_file_minidump != undefined && result.files.upload_file_minidump > 1024;
 
         this._addEvent({
-          data: event,
-          namespacedData: namespaced,
-          dumpFile: dump_file,
-          appId: ctx.params.id,
-          sentryKey: keyMatch[1],
+          eventData: event,
+          namespaced,
+          dump_file,
+          id: ctx.params.id,
+          sentry_key: keyMatch[1],
           method: 'minidump',
         });
 
@@ -128,11 +130,11 @@ export class TestServer {
       const event = JSON.parse(ctx.request.body as string);
 
       this._addEvent({
-        data: event,
-        appId: ctx.params.id,
-        sentryKey: keyMatch[1],
+        eventData: event,
+        id: ctx.params.id,
+        sentry_key: keyMatch[1],
         method: 'store',
-        dumpFile: false,
+        dump_file: false,
       });
 
       ctx.headers['Content-Type'] = 'text/plain; charset=utf-8';